refactored page move to homepage
