--- app/refactor/page.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import React from 'react';
-import ThreadedDocument from '@/components/threaded-document';
-
-export default function Home() {
-  return (
-    <main>
-      <ThreadedDocument />
-    </main>
-  );
-}
