+++ components/hooks/use-messages.ts
@@ -3,6 +3,7 @@ import { Message } from '../types';
 
 export function useMessages() {
     const [selectedMessage, setSelectedMessage] = useState<string | null>(null);
+    const [replyingTo, setReplyingTo] = useState<string | null>(null);
     const [editingMessage, setEditingMessage] = useState<string | null>(null);
     const [editingContent, setEditingContent] = useState("");
     const [clipboardMessage, setClipboardMessage] = useState<{
@@ -11,15 +12,20 @@ export function useMessages() {
         sourceThreadId: string | null;
         originalMessageId: string | null;
     } | null>(null);
+    const [glowingMessageId, setGlowingMessageId] = useState<string | null>(null);
 
     return {
         selectedMessage,
         setSelectedMessage,
+        replyingTo,
+        setReplyingTo,
         editingMessage,
         setEditingMessage,
         editingContent,
         setEditingContent,
         clipboardMessage,
-        setClipboardMessage
+        setClipboardMessage,
+        glowingMessageId,
+        setGlowingMessageId,
     };
 }
